PROF-109 [feat]: 댓글 생성시 유저 검증 로직 추가

+++ src/main/java/com/profitkey/stock/dto/response/community/CommunityResponse.java
@@ -1,7 +1,9 @@
 package com.profitkey.stock.dto.response.community;
 
-import com.profitkey.stock.entity.Community;
 import java.time.LocalDateTime;
+
+import com.profitkey.stock.entity.Community;
+
 import lombok.Builder;
 import lombok.Getter;
 
@@ -9,7 +11,7 @@ import lombok.Getter;
 @Builder
 public class CommunityResponse {
 	private final String id;
-	private final Integer writerId;
+	private final Long writerId;
 	private final String parentId;
 	private final String content;
 	private final LocalDateTime createdAt;
